agregando README.MD y mas cosillas

+++ src/main/java/com/example/apiDisney/service/impl/MovieServiceImpl.java
@@ -5,11 +5,17 @@ import com.example.apiDisney.model.GenderEntity;
 import com.example.apiDisney.model.MovieEntity;
 import com.example.apiDisney.repository.MovieRepository;
 import com.example.apiDisney.service.MovieService;
+import com.example.apiDisney.service.exception.CustomException;
+import javassist.bytecode.ExceptionsAttribute;
+import org.hibernate.exception.DataException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class MovieServiceImpl implements MovieService {
@@ -23,45 +29,55 @@ public class MovieServiceImpl implements MovieService {
     @Autowired
     private GenderServiceImpl genderService;
 
-    public MovieEntity save(MovieEntity movieEntity) throws Exception {
-        boolean exists = movieRepository.getByTitle(movieEntity.getTitle())
-                .stream().map(movie ->
-                        movie.getCreationdate().equals(movieEntity.getCreationdate()))
-                .reduce(false,(x,y) -> x || y);
-        if(exists){
-            throw new Exception("Movie with title and  creation date exists in database");
+    public MovieEntity save(MovieEntity movieEntity) throws CustomException,DataIntegrityViolationException{
+        if(movieRepository.existsByTitleAndCreationdate1(movieEntity.getTitle(),movieEntity.getCreationdate())){
+            throw new CustomException("Movie with title and  creation date exists in database");
         }
-        try {
+       try {
             return movieRepository.save(movieEntity);
-        }catch (Exception ex){
-            throw new Exception("Enter corrects properties");
+        }catch (DataIntegrityViolationException ex) {
+           throw new DataIntegrityViolationException("Enter corrects properties");
+       }
+    }
+
+    @Override
+    public void update(Long id, MovieEntity movieEntity) throws EmptyResultDataAccessException, DataIntegrityViolationException, CustomException{
+        movieEntity.setId(findById(id).getId());
+
+        if(movieRepository.existsByTitleAndCreationdate1(movieEntity.getTitle(),movieEntity.getCreationdate())) {
+            //Se testea en trySaveMovieWithTitleAndCreationDateExists()
+            throw new CustomException("Movie with title and  creation date exists in database");
+        }else{
+            movieRepository.save(movieEntity);
         }
     }
 
-    public MovieEntity update(MovieEntity movieEntity) throws Exception{
-        return movieRepository.save(movieEntity);
+    public MovieEntity findById(Long id) {
+        return movieRepository.findById(id).orElseThrow(() -> new EmptyResultDataAccessException("Entity with id "+id+" not found.",0));
     }
 
     public List<MovieEntity> getMovies(){
         return movieRepository.findAll();
     }
-
-    //No puede eliminarse
-    //El id es cualquier cosa
-    public void deleteById(Long id) throws Exception {
+//EmptyResultDataAccessException ,
+    public void deleteById(Long id) throws CustomException,NoSuchElementException{
         try {
-            movieRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException ex) {
-            throw new EmptyResultDataAccessException("Movie with id " + id + " not found in database", 0);
+            if (movieRepository.findById(id).get().getCharacterEntities().isEmpty()) {
+                movieRepository.deleteById(id);
+            } else {
+                throw new CustomException("Movie has charecters asigned,try deleted first the characters for delete this movie");
+            }
+            //testeado en tryDeleteByIdNotExisting()
+        }catch (NoSuchElementException ex){
+            throw new NoSuchElementException("Movie with id " + id + " not found in database");
         }
     }
-
 
     public void deleteAll() {
         movieRepository.deleteAll();
     }
 
-    public List<MovieEntity> getByName(String name) {
+    public List<MovieEntity> getByTitle(String name) {
         return movieRepository.getByTitle(name);
     }
 
@@ -77,39 +93,53 @@ public class MovieServiceImpl implements MovieService {
         return movieRepository.findByOrderByCreationdateAsc();
     }
 
-    public MovieEntity setCharacterInMovie(Long movie_id, Long character_id) throws Exception {
+    public MovieEntity addCharacterInMovie(Long movie_id, Long character_id) throws InvalidDataAccessApiUsageException,EmptyResultDataAccessException, CustomException{
         try{
             MovieEntity movieEntity = this.findById(movie_id);
             CharacterEntity characterEntity = characterService.findById(character_id);
             if (movieEntity.getCharacterEntities().contains(characterEntity)){
-                throw new Exception("This Movie contains character "+characterEntity.getName());
+                throw new InvalidDataAccessApiUsageException("This Movie contains character "+characterEntity.getName());
             }else{
                 movieEntity.addCharacter(characterEntity);
+                characterService.update(characterEntity.getId(),characterEntity);
                 return movieRepository.save(movieEntity);
             }
-        }catch (Exception ex){
-            throw new Exception("Enter corrects id's for movie or Gdender");
+        }catch (EmptyResultDataAccessException ex){
+            throw new EmptyResultDataAccessException("Enter corrects id for movie or Gdender",0);
         }
     }
 
-    public MovieEntity setGenderInMovie(Long movie_id, Long gender_id) throws Exception{
-
+    public MovieEntity addGenderInMovie(Long movie_id, Long gender_id) throws InvalidDataAccessApiUsageException, EmptyResultDataAccessException{
         try{
             MovieEntity movieEntity = this.findById(movie_id);
             GenderEntity genderEntity = genderService.findById(gender_id);
             if (movieEntity.getGenderEntities().contains(genderEntity)){
-                throw new Exception("This Movie contains gender "+genderEntity.getName());
+                /*This case is tested "tryAddAGenderIncludeInMovie()"*/
+                throw new InvalidDataAccessApiUsageException("This Movie contains gender "+genderEntity.getName());
             }else{
                 movieEntity.addGender(genderEntity);
                 return movieRepository.save(movieEntity);
             }
-        }catch (Exception ex){
-            throw new Exception("Enter corrects id's for movie or Gdender");
+        }catch (EmptyResultDataAccessException ex){
+            throw new EmptyResultDataAccessException("Enter corrects id's for movie or Gdender",0);
         }
     }
 
-    public MovieEntity findById(Long id) throws Exception{
-        return movieRepository.findById(id).orElseThrow(() -> new Exception("Not Found"));
+    @Override
+    public MovieEntity removeCharacter(Long movieId, Long characterId) throws CustomException{
+        try{
+            MovieEntity movieEntity = this.findById(movieId);
+            CharacterEntity characterEntity = characterService.findById(characterId);
+            if (movieEntity.getCharacterEntities().contains(characterEntity)){
+                throw new CustomException("This Movie NO contains character "+characterEntity.getName());
+            }else{
+                movieEntity.removeCharacter(characterEntity);
+                characterService.save(characterEntity);
+                return movieRepository.save(movieEntity);
+            }
+        }catch (EmptyResultDataAccessException ex){
+            throw new EmptyResultDataAccessException("Enter corrects id for movie or Gdender",0);
+        }
     }
 
 }
